GetBrand

+++ src/Infrastructure/Food/Persistence/DoctrineBrandsCatalog.php
@@ -6,6 +6,8 @@ use App\Domain\Food\Entity\Brand;
 use App\Domain\Food\Exception\BrandNotFoundException;
 use App\Domain\Food\Repository\BrandsCatalog;
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Exception\ORMException;
+use Doctrine\ORM\ORMInvalidArgumentException;
 
 class DoctrineBrandsCatalog implements BrandsCatalog
 {
@@ -20,10 +22,13 @@ class DoctrineBrandsCatalog implements BrandsCatalog
 
     public function get(string $id): Brand
     {
-        $brand = $this->entityManager->find(Brand::class, $id);
-
-        if (null === $brand) {
-            throw new BrandNotFoundException($id);
+        try {
+            $brand = $this->entityManager->find(Brand::class, $id);
+            if (null === $brand) {
+                throw new BrandNotFoundException($id);
+            }
+        } catch (ORMInvalidArgumentException|ORMException $exception) {
+            throw new \LogicException(sprintf('Impossible to retrieve brand with id %s', $id));
         }
 
         return $brand;
@@ -31,8 +36,18 @@ class DoctrineBrandsCatalog implements BrandsCatalog
 
     public function remove(string $id): void
     {
-        $brand = $this->get($id);
+        try {
+            $brand = $this->entityManager->getReference(
+                Brand::class,
+                $id,
+            );
+            if (null === $brand) {
+                throw new BrandNotFoundException($id);
+            }
 
-        $this->entityManager->remove($brand);
+            $this->entityManager->remove($brand);
+        } catch (ORMInvalidArgumentException|ORMException $exception) {
+            throw new \LogicException(sprintf('Impossible to remove brand with id %s', $id));
+        }
     }
 }
